pyproject.toml
[project]
name = "prayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "alembic>=1.14.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asyncpg>=0.30.0",
    "attrs==25.1.0",
    "authlib==1.3.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cryptography==44.0.1",
    "dataclasses-json==0.6.7",
    "deepgram-sdk>=3.10.0",
    "fastapi>=0.115.8",
    "filelock==3.17.0",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "grpcio-health-checking==1.70.0",
    "grpcio-tools==1.70.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.28.1",
    "idna==3.10",
    "jinja2==3.1.5",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.19",
    "langchain-community==0.3.17",
    "langchain-core==0.3.35",
    "langchain-openai>=0.3.6",
    "langchain-text-splitters==0.3.6",
    "langchain-weaviate==0.0.4",
    "langsmith==0.3.8",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "orjson==3.10.15",
    "packaging==24.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow==11.1.0",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "psycopg2-binary>=2.9.10",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pyjwt>=2.10.1",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.20",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "safetensors==0.5.2",
    "scipy==1.15.2",
    "setuptools==75.8.0",
    "simsimd==6.2.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "urllib3==2.3.0",
    "uvicorn>=0.34.0",
    "validators==0.34.0",
    "weaviate-client==4.10.4",
    "yarl==1.18.3",
    "zstandard==0.23.0",
]
